Changed abstract list to interface.
