+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/persistence/GsonExpenseClaimListPersister.java
@@ -2,7 +2,6 @@ package ca.ualberta.cs.shinyexpensetracker.persistence;
 
 import java.io.IOException;
 
-import ca.ualberta.cs.shinyexpensetracker.models.ConcreteExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaimList;
 
 import com.google.gson.Gson;
@@ -32,7 +31,7 @@ public class GsonExpenseClaimListPersister implements IExpenseClaimListPersister
 	public ExpenseClaimList loadExpenseClaims() throws IOException {
 		String travelClaimsListData = persistenceStrategy.load();
 		if (travelClaimsListData.equals("")) {
-			return new ConcreteExpenseClaimList();
+			return new ExpenseClaimList();
 		} else {
 			return gson.fromJson(travelClaimsListData, ExpenseClaimList.class);
 		}
