+++ ShinyExpenseTrackerTests/src/ca/ualberta/cs/shinyexpensetracker/test/EditExpenseClaimTests.java
@@ -14,7 +14,6 @@ import ca.ualberta.cs.shinyexpensetracker.R;
 import ca.ualberta.cs.shinyexpensetracker.activities.ExpenseClaimActivity;
 import ca.ualberta.cs.shinyexpensetracker.framework.Application;
 import ca.ualberta.cs.shinyexpensetracker.framework.ExpenseClaimController;
-import ca.ualberta.cs.shinyexpensetracker.models.ConcreteExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.test.mocks.MockExpenseClaimListPersister;
@@ -52,7 +51,7 @@ public class EditExpenseClaimTests extends ActivityInstrumentationTestCase2<Expe
 	protected void setUp() throws Exception {
 		super.setUp();
 
-		list = new ConcreteExpenseClaimList();
+		list = new ExpenseClaimList();
 		claim = getStartingClaim();
 
 		list.addClaim(claim);
