+++ ShinyExpenseTrackerTests/src/ca/ualberta/cs/shinyexpensetracker/test/EditExpenseItemTests.java
@@ -21,7 +21,6 @@ import ca.ualberta.cs.shinyexpensetracker.R;
 import ca.ualberta.cs.shinyexpensetracker.activities.ExpenseItemActivity;
 import ca.ualberta.cs.shinyexpensetracker.framework.Application;
 import ca.ualberta.cs.shinyexpensetracker.framework.ExpenseClaimController;
-import ca.ualberta.cs.shinyexpensetracker.models.ConcreteExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseItem;
@@ -66,7 +65,7 @@ public class EditExpenseItemTests extends ActivityInstrumentationTestCase2<Expen
 	protected void setUp() throws Exception {
 		super.setUp();
 
-		ExpenseClaimList claimList = new ConcreteExpenseClaimList();
+		ExpenseClaimList claimList = new ExpenseClaimList();
 		persister = new MockExpenseClaimListPersister(claimList);
 		controller = new ExpenseClaimController(persister);
 		Application.setExpenseClaimController(controller);
