+++ ShinyExpenseTrackerTests/src/ca/ualberta/cs/shinyexpensetracker/test/TestReceiptView.java
@@ -15,7 +15,6 @@ import ca.ualberta.cs.shinyexpensetracker.R;
 import ca.ualberta.cs.shinyexpensetracker.activities.ReceiptViewActivity;
 import ca.ualberta.cs.shinyexpensetracker.framework.Application;
 import ca.ualberta.cs.shinyexpensetracker.framework.ExpenseClaimController;
-import ca.ualberta.cs.shinyexpensetracker.models.ConcreteExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaimList;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseItem;
@@ -46,7 +45,7 @@ public class TestReceiptView extends ActivityInstrumentationTestCase2<ReceiptVie
 	protected void setUp() throws Exception {
 		/* Set a test item in a claim list and ask the activity to fetch it */
 		super.setUp();
-		ExpenseClaimList claimList = new ConcreteExpenseClaimList();
+		ExpenseClaimList claimList = new ExpenseClaimList();
 		controller = new ExpenseClaimController(new MockExpenseClaimListPersister(claimList));
 		Application.setExpenseClaimController(controller);
